third_party/blink/web_tests/html/details_summary/details-open-toggle-event-expected.txt

Tests that a 'toggle' event is fired asynchronously whenever the open attribute is added to or removed from a details element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS details.ontoggle.__proto__ is Function.prototype
PASS details.open is false
PASS toggleEventCount is 0
PASS toggleEventCountFromHTMLAttribute is 0
PASS toggleEventCount is 0
PASS toggleEventCountFromHTMLAttribute is 0
PASS testEvent.__proto__ is ToggleEvent.prototype
PASS testEvent.type is "toggle"
PASS testEvent.__proto__ is ToggleEvent.prototype
PASS testEvent.type is "toggle"
PASS toggleEventCount is 1
PASS toggleEventCountFromHTMLAttribute is 1
PASS details.open is true
PASS testEvent.__proto__ is ToggleEvent.prototype
PASS testEvent.type is "toggle"
PASS testEvent.__proto__ is ToggleEvent.prototype
PASS testEvent.type is "toggle"
PASS toggleEventCount is 2
PASS toggleEventCountFromHTMLAttribute is 2
PASS details.open is false
PASS successfullyParsed is true

TEST COMPLETE

details
